Update R13_FIO08_J.java
compliant code

+++ R13_FIO08_J.java
@@ -1,6 +1,6 @@
 FileInputStream in = new FileInputStream("file");
-byte data;
-while ((data = (byte) in.read()) != -1) {
+int data;
+while ((data = in.read()) != -1) {
     // Process data
 }
 in.close();
